+++ inst/include/rodeint/util_ublas.hpp
@@ -13,7 +13,7 @@ ublas_vector_to_r(const boost::numeric::ublas::vector<T>& v) {
 template <typename T>
 boost::numeric::ublas::vector<T>
 r_vector_to_ublas(const Rcpp::NumericVector& v) {
-  boost::numeric::ublas::vector<T> ret(v.size());
+  boost::numeric::ublas::vector<T> ret(static_cast<size_t>(v.size()));
   std::copy(v.begin(), v.end(), ret.begin());
   return ret;
 }
@@ -22,7 +22,7 @@ template <typename T>
 Rcpp::NumericMatrix
 ublas_matrix_to_r(const boost::numeric::ublas::matrix<T>& m) {
   const size_t nr = m.size1(), nc = m.size2();
-  Rcpp::NumericMatrix ret(nr, nc);
+  Rcpp::NumericMatrix ret(static_cast<int>(nr), static_cast<int>(nc));
   for (size_t i = 0; i < nr; ++i) {
     for (size_t j = 0; j < nc; ++j) {
       ret(i, j) = m(i, j);
@@ -34,7 +34,9 @@ ublas_matrix_to_r(const boost::numeric::ublas::matrix<T>& m) {
 template <typename T>
 boost::numeric::ublas::matrix<T>
 r_matrix_to_ublas(const Rcpp::NumericMatrix& m) {
-  const size_t nr = m.rows(), nc = m.cols();
+  const size_t
+    nr = static_cast<size_t>(m.rows()),
+    nc = static_cast<size_t>(m.cols());
   boost::numeric::ublas::matrix<T> ret(nr, nc);
   for (size_t i = 0; i < nr; ++i) {
     for (size_t j = 0; j < nc; ++j) {
